+++ account_invoice_import/__init__.py
@@ -1,2 +1,3 @@
+from . import components
 from . import models
 from . import wizard
