+++ account_invoice_import/components/__init__.py
@@ -1,0 +1,1 @@
+from . import edi_input_process_business_document_account_invoice
