+++ account_invoice_import/models/__init__.py
@@ -3,3 +3,4 @@ from . import company
 from . import res_config_settings
 from . import account_invoice_import_config
 from . import account_move
+from . import edi_backend
