+++ account_invoice_import/views/account_invoice_import_config.xml
@@ -9,49 +9,45 @@
         <field name="model">account.invoice.import.config</field>
         <field name="arch" type="xml">
             <form string="Invoice Import Configuration">
-                <div class="oe_button_box" name="button_box">
-                    <button
-                        name="toggle_active"
-                        type="object"
-                        class="oe_stat_button"
-                        icon="fa-archive"
-                    >
+                <sheet>
+                    <widget
+                        name="web_ribbon"
+                        title="Archived"
+                        bg_color="bg-danger"
+                        attrs="{'invisible': [('active', '=', True)]}"
+                    />
+                    <div class="oe_button_box" name="button_box" />
+                    <group name="main">
+                        <field name="active" invisible="1" />
+                        <field name="name" />
+                        <field name="partner_id" />
+                        <field name="company_id" groups="base.group_multi_company" />
+                    </group>
+                    <group string="Accounting Parameters" name="accounting">
+                        <field name="invoice_line_method" />
                         <field
-                            name="active"
-                            widget="boolean_button"
-                            options='{"terminology": "archive"}'
+                            name="account_id"
+                            attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_no_product', 'nline_no_product'))], 'required': [('invoice_line_method', 'in', ('1line_no_product', 'nline_no_product'))]}"
                         />
-                    </button>
-                </div>
-                <group name="main">
-                    <field name="name" />
-                    <field name="partner_id" />
-                    <field name="company_id" groups="base.group_multi_company" />
-                </group>
-                <group string="Accounting Parameters" name="accounting">
-                    <field name="invoice_line_method" />
-                    <field
-                        name="account_id"
-                        attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_no_product', 'nline_no_product'))], 'required': [('invoice_line_method', 'in', ('1line_no_product', 'nline_no_product'))]}"
-                    />
-                    <field
-                        name="tax_ids"
-                        widget="many2many_tags"
-                        attrs="{'invisible': [('invoice_line_method', '!=', '1line_no_product')]}"
-                    />
-                    <field
-                        name="static_product_id"
-                        attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_static_product', 'nline_static_product'))], 'required': [('invoice_line_method', 'in', ('1line_static_product', 'nline_static_product'))]}"
-                    />
-                    <field
-                        name="account_analytic_id"
-                        groups="analytic.group_analytic_accounting"
-                    />
-                    <field
-                        name="label"
-                        attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_no_product', '1line_static_product'))]}"
-                    />
-                </group>
+                        <field
+                            name="tax_ids"
+                            widget="many2many_tags"
+                            attrs="{'invisible': [('invoice_line_method', '!=', '1line_no_product')]}"
+                        />
+                        <field
+                            name="static_product_id"
+                            attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_static_product', 'nline_static_product'))], 'required': [('invoice_line_method', 'in', ('1line_static_product', 'nline_static_product'))]}"
+                        />
+                        <field
+                            name="account_analytic_id"
+                            groups="analytic.group_analytic_accounting"
+                        />
+                        <field
+                            name="label"
+                            attrs="{'invisible': [('invoice_line_method', 'not in', ('1line_no_product', '1line_static_product'))]}"
+                        />
+                    </group>
+                </sheet>
             </form>
         </field>
     </record>
